dbt_run_completed_function/requirements.txt
azure-functions==1.21.3
azure-mgmt-datafactory==9.1.0
azure-identity==1.19.0
opencensus-ext-azure==1.1.14
# https://github.com/Azure/azure-functions-python-worker/issues/1050
protobuf >=5.27.4, ==5.*
pyyaml==6.0.2
